refactor io_error, file, tcp
