+++ Cargo.toml
@@ -4,7 +4,10 @@ members = [
     "crates/roc_command",
     "crates/roc_host",
     "crates/roc_host_lib",
+    "crates/roc_file",
     "crates/roc_host_bin",
+    "crates/roc_tcp",
+    "crates/roc_io_error",
 ]
 
 [workspace.package]
@@ -25,3 +28,19 @@ codegen-units = 1
 roc_std = { git = "https://github.com/roc-lang/roc.git" }
 roc_std_heap = { git = "https://github.com/roc-lang/roc.git" }
 roc_command = { path = "crates/roc_command" }
+roc_file = { path = "crates/roc_file" }
+roc_host = { path = "crates/roc_host" }
+roc_tcp = { path = "crates/roc_tcp" }
+roc_io_error = { path = "crates/roc_io_error" }
+
+memchr = "=2.7.4"
+hyper = { version = "=0.14.27", default-features = false, features = [
+    "http1",
+    "client",
+] }
+hyper-rustls = { version = "=0.24.2", default-features = false, features = [
+    "http1",
+    "tls12",
+    "native-tokio",
+] }
+tokio = { version = "=1.31.0", default-features = false }
