+++ crates/roc_file/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "roc_file"
+description = "Common functionality for Roc to interface with std::io"
+
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+version.workspace = true
+
+[dependencies]
+roc_std.workspace = true
+roc_std_heap.workspace = true
+roc_io_error.workspace = true
+memchr.workspace = true
