+++ crates/roc_host/Cargo.toml
@@ -16,18 +16,14 @@ path = "src/lib.rs"
 roc_std.workspace = true
 roc_std_heap.workspace = true
 roc_command.workspace = true
+roc_file.workspace = true
+roc_io_error.workspace = true
+roc_tcp.workspace = true
 libc = "=0.2.155"
 backtrace = "=0.3.69"
-hyper = { version = "=0.14.27", default-features = false, features = [
-    "http1",
-    "client",
-] }
-hyper-rustls = { version = "=0.24.2", default-features = false, features = [
-    "http1",
-    "tls12",
-    "native-tokio",
-] }
-tokio = { version = "=1.31.0", default-features = false }
+hyper.workspace = true
+hyper-rustls.workspace = true
+tokio.workspace = true
 crossterm = "=0.27.0"
 memmap2 = "=0.9.4"
 memchr = "=2.7.4"
