+++ crates/roc_host_bin/Cargo.toml
@@ -12,4 +12,4 @@ path = "src/main.rs"
 
 [dependencies]
 roc_std.workspace = true
-roc_host = { path = "../roc_host" }
+roc_host.workspace = true
