+++ crates/roc_tcp/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "roc_tcp"
+description = "Common functionality for Roc to interface with std::net::tcp"
+
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+version.workspace = true
+
+[dependencies]
+roc_std.workspace = true
+roc_std_heap.workspace = true
+roc_io_error.workspace = true
+roc_file.workspace = true
+memchr.workspace = true
+hyper.workspace = true
+hyper-rustls.workspace = true
+tokio.workspace = true
